Create reservation rate report components
Create component ReservationRateReportModal. This modal lies
in the /reservations page and can be unhidden/shown in
the dropdown of <Lataa Raportti> -> Varausasteraportti.

This modal will take the basic props the modal itself requires
along with units and the action which downloads the report.

Within the modal there are fields where the user can select/
input data: Units, Date range and Time range. Unit seletion field
is a Typeahead component which is also a new package. Lastly
there is the download button which when clicked, will call
the api action to download the report.