Reservation rate excel report (#338)
* Create Reservation rate report api action

Create the action for fetching the reservation
rate report excel.

Also add extensions for filenames for all report actions
and modify the downloadReport function to not append
.docx to filename anymore. This had to be done because
all previous reports are .docx but this new one is xlsx.

* Conditional datepicker rendering in DateTimeRange

Conditionally render datepicker in DateTimeRange
component. Defaults to always rendering it unless
specifying not to render it in controlprops.

The Time range part of the DateTimeRange component
is nice and useful, so now basically you are able
to just render that part.

* Reservation Rate Report actions, reducers, selectors

Create actions reducers and selectors for Modal.

* Create reservation rate report components

Create component ReservationRateReportModal. This modal lies
in the /reservations page and can be unhidden/shown in
the dropdown of <Lataa Raportti> -> Varausasteraportti.

This modal will take the basic props the modal itself requires
along with units and the action which downloads the report.

Within the modal there are fields where the user can select/
input data: Units, Date range and Time range. Unit seletion field
is a Typeahead component which is also a new package. Lastly
there is the download button which when clicked, will call
the api action to download the report.

* Fix errors

Errors were mostly due to incorrect proptyping

* Project conf

Install node-sass that is compatible with
node version 14.X.X

Add a css loader for testing.

Co-authored-by: Kevin Seestrand <kevin.seestrand@anders.fi>